service에서 바로 하기
