+++ lib/app/data/services/journal_service/journal_repository.dart
@@ -27,21 +27,4 @@ class JournalRepository {
 
     return journalList;
   }
-
-  // get Journal -> 읽은 후 type으로 뭔지 판단
-  Future<List<Journal?>> getJournalListWithinThePeriod(
-      DateTime from, DateTime to) async {
-    // from 보다 커야하고 to 보다 작아야 한다!
-    List<Map<String, dynamic>> journalJsonList =
-        await _dbProvider.readJournalList();
-    List<Journal?> journalList = journalJsonList.map((journalJson) {
-      var time = Journal.fromJson(journalJson).date;
-      if (from.difference(time).inDays <= 0 &&
-          to.difference(time).inDays >= 0) {
-        return Journal.fromJson(journalJson);
-      }
-    }).toList();
-
-    return journalList;
-  }
 }
