+++ lib/app/data/services/journal_service/journal_service.dart
@@ -1,3 +1,5 @@
+import 'dart:developer';
+
 import 'package:alcohol_free/app/data/models/drinking_journal.dart';
 import 'package:alcohol_free/app/data/models/sobriety_journal.dart';
 import 'package:alcohol_free/app/data/services/journal_service/journal_repository.dart';
@@ -31,10 +33,32 @@ class JournalService extends GetxService {
     return journal;
   }
 
+  /*
+    // get Journal -> 읽은 후 type으로 뭔지 판단
   Future<List<Journal?>> getJournalListWithinThePeriod(
       DateTime from, DateTime to) async {
-    var journalListWithinThePeriod =
-        await _journalRepository.getJournalListWithinThePeriod(from, to);
+    // from 보다 커야하고 to 보다 작아야 한다!
+    List<Map<String, dynamic>> journalJsonList =
+        await _dbProvider.readJournalList();
+    List<Journal?> journalList = journalJsonList.map((journalJson) {
+      var time = Journal.fromJson(journalJson).date;
+      if (from.difference(time).inDays <= 0 &&
+          to.difference(time).inDays >= 0) {
+        return Journal.fromJson(journalJson);
+      }
+    }).toList();
+
+    return journalList;
+  }
+   */
+  List<Journal?> getJournalListWithinThePeriod(DateTime from, DateTime to) {
+    //log(_journalList.toString());
+    List<Journal?> journalListWithinThePeriod = _journalList.map((journalJson) {
+      var time = journalJson.date;
+      if (from.compareTo(time) <= 0 && to.compareTo(time) >= 0) {
+        return journalJson;
+      }
+    }).toList();
     return journalListWithinThePeriod;
   }
 }
